requirements.txt
Flask
python-dotenv
mindsdb-sdk
Flask-Cors
